Bug fixed, jQuery intergration started
Found a major issue on line 5 where "ASB.js" was inputed as "ASB,js". This has been corrected. jQuery intergration has been started but is still no working. Version number is in a comment under the closing footer tag, please update it as you update the code to help keep track of the current version. Ver. 0.02